+++ Kronko/ColorGauss.h
@@ -5,6 +5,6 @@ class ColorGauss :
 {
 public:
     using ColorPicker::ColorPicker;
-    Vec3i getColorV(Mat& img, Point p = Point(-1, -1), double radius = -1.0) override;
+    Vec3i getColorV(Mat& img, Point p = Point(-1, -1), int radius = -1.0) override;
 };
